[Mem-mgmt]Created class for mem-mgmt
